Add deploy-csb pipeline, ported from deploy-cf; add docproxy terraform
Compliment to PR: https://github.com/cloud-gov/deploy-cf/pull/944

Reiterating the reasoning from that commit:

The set-self step on deploy-cf blocks when the deploy-cf-* jobs (or any other jobs in the pipeline) are running. This means that the deploy-apps- jobs may have to wait hours to get updated with changes merged to main. Moving the jobs to a separate pipeline will fix this issue. This will also focus each pipeline on a single responsibility. Lastly, developers will be able to set the deploy-csb pipeline to watch a topic branch for faster iteration without affecting CF deployments.

This also includes some code for deploying the docproxy.