Add support for redirecting requests when HOST does not match
Due to the path route cloudfoundry_route.csb_docs, requests made to the CSB host, but with the /docs path, will be sent to the docproxy. This is desired so customers cannot see the unformatted documentation page. However, when that happens, the docproxy tries loading assets using relative paths based on the HOST header. Since the host is still the CSB's URL, those requests fail. This change handles this more cleanly: Requests send to the docproxy, but with the broker's HOST, are redirected.