Fix panic on SNS subscription requests and test the handler
Calling sns.Client.Options().BaseEndpoint was not the correct way to get the base endpoint if that endpoint was automatically populated -- only if it was set manually. The correct method is now used in main.go: Use NewDefaultEndpointResolver in the service package (e.g `service/sns`). As of this commit the code passes the test locally, using a local server.