Improve CSB deployment failure message
We only keep 2 versions of docker images in each ECR registry, the two most recent. We delete older versions automatically. The CSB and Helper reference images by SHA, not floating tag. If this pipeline does not run, eventually the running app's SHA will reference a version that was removed from the registry. If the app restarts and tries to pull the image, it will fail, resulting in an outage.